Add libevent hooks.
Use libevent to receive packets from both interfaces that we bind to.
Right now it is very ugly with lots of code-smells (copy-pasted code)
but it should get resolved soon. Also at this point only ARP packets
seem to work correctly and the frame-length that is used is completely
wrong when we modify and send the packets. This should also change soon.