Map IP addresses to MAC addresses.
Each IP address for packets coming in is now mapped to exactly one MAC
address (per IP address, if the previous sentence is not too clear).
This is done using a MAC to IP mapping table (using a simple array for
now). the last digit(s) of the IP address is used as the key to look up
the MAC address in the array.
Example for key selection:
    IP        |    Key
+-------------+----------+
|10.10.3.2    |     2    |
|10.10.3.3    |     3    |
|10.10.3.29   |    29    |

and so on.